[ir-xenon-tycoon] Delete cached model data before moving (#4846)
## Summary
This change:

1. Updates the downloader to use a unique UUID-based temporary URL for
downloads.
2. When a model fails to load, proactively attempt to delete it.
3. When caching a model, delete the previous entry.

## Motivation
ir-xenon-tycoon is showing sporadic failures during identity
verification reporting NSCocoaErrorDomain code 516 when attempting to
start scanning. This error is caused when a file already exists at the
destination trying to create or move. Based on the time at which it
happens, it appears that this occurs during model loading. There are
several places where we are moving into the new destination without
checking if contents already exists -- and the file downloader is using
the last path component to generate a temporary URL.

## Testing
Manually validated the flow with new installations and existing
installations using the test app.

## Changelog
[TBD]