Refactor Account Picker text strings for better localization (#4855)
## Summary
Updates some text on the native account picker so that they're more
likely to be localized correctly.

## Motivation
We've received feedback that the templating of some of the strings on
the page may result in incorrect localization in certain languages due
to compounded nouns and adjective declension.

## Testing
No functional changes and visually inspected with previews.

---------

Co-authored-by: Mat Schmid <mats@stripe.com>