Render Link payment methods in SPM screens (#4861)
## Summary
<!-- Simple summary of what was changed. -->

This pull request introduces the ability to show saved payment methods
of type `link` in the list of saved payment methods. This functionality
is available to merchants that integrate with CustomerSession.

<details><summary>Screen recording</summary>
<p>


https://github.com/user-attachments/assets/dbb5d70c-de6c-45ce-9cda-b311721e4477

</p>
</details> 

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

Allowing Link payment methods in CustomerSession.

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->

Nothing yet. Will eventually remove the feature flag and then add a
release note.