Always use "New Card" copy when there is a saved card in Vertical Payment Sheet (#4841)
## Summary
Previously, if there were multiple saved payment methods, we would only
show the "New card" copy (as opposed to "Card") if the first saved
payment method was card. Now, we will always use the "New card" copy if
there is an existing saved card.

## Motivation
https://jira.corp.stripe.com/browse/MOBILESDK-2831

## Testing
- added a unit test that confirms that "New card" is used when there is
a saved card but not as the first saved payment method
- added a unit test that confirms that "Card" is used if there is not a
saved card