[LINK_MOBILE-162] Link signup screen button tweaks (#4878)
## Summary

Adds a disabled `Log in or sign up` button when just the email field is
shown. The remaining logic doesn't change - if a known email is
detected, we automatically go to the OTP verification screen, otherwise
we expand the other fields and the button switches to `Agree and
continue`.

## Motivation

 [LINK_MOBILE-162](https://jira.corp.stripe.com/browse/LINK_MOBILE-162)

## Testing


https://github.com/user-attachments/assets/63801086-7158-4039-a306-1e91cb5b5422

## Changelog

N/a