Fix 3DS2 Whitelisting Encoding Issue (#4923)
## Summary
<!-- Simple summary of what was changed. -->
Fixes an issue where we send a whitelisting value, even if no text is
present. Some ACSs seem to validate that this value should *NOT* exist
when the whitelisting text is not present.

## Motivation
An ACS reported failures when trying to validate challenges that did not
contain whitelist text.

## Testing
Using the PaymentSheet Example to test cards that did not contain
whitelist text to ensure that the field was no longer sent, and then
validating that it was still sent when text did exist.

---------

Co-authored-by: tjclawson-stripe <163896025+tjclawson-stripe@users.noreply.github.com>