Send customer ephemeral key to backend (#4897)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds a `legacy_customer_ephemeral_key` parameter to
the Elements session request. This parameter helps us determine if
inline signup can be enabled. The related backend work can be found
[here](https://git.corp.stripe.com/stripe-internal/pay-server/pull/1084214).

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->