Support adding bank accounts in Link (#4933)
## Summary

This wires up the changes introduced in the [upstream
branch](https://github.com/stripe/stripe-ios/pull/4932) for adding bank
accounts to Link. It uses the Financial Connections SDK if available,
and otherwise uses FC Lite. The general flow is;

1. Create a Link Account Session
2. Launch Financial Connections using the client secret we get in step 1
3. Create the payment details using the account ID we get in step 2
4. Refresh the Link payment method list

## Motivation

Part 2 of 2 of
https://docs.google.com/document/d/1S7OyhUuqi7ESFO4rZDt71exHCyyZwGBWKQPr7m9LLJs/edit?usp=sharing

## Testing

Native SDK:


https://github.com/user-attachments/assets/6766aed1-21d2-4a5b-b963-52f0b0f0055f

Web SDK:


https://github.com/user-attachments/assets/13b17538-c6af-4c45-843c-26561e22da24

FC Lite


https://github.com/user-attachments/assets/3d63f0ec-9905-44e7-9635-9ce5c2d9887c

## Changelog

N/a