Add alternative Link flow in FlowController (#4911)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds an alternative Link flow to FlowController. All
changes only apply to the native Link experience and are hidden behind a
local feature flag.

The main changes include:
- Immediately show the Link UI upon selection of Link in FlowController
(either via the wallet button or the primary button)
- Open the Link UI instead of FlowController when the currently selected
payment method is a Link payment method

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->