Appearance API formInsets (#4941)
## Summary
<!-- Simple summary of what was changed. -->
Added formInsets to Appearance API

Added vars in an extension of PaymentSheet.Appearance for ease of use
allFormInsets converts formInsets to NSDirectionalEdgeInsets using all 4
values
topFormInsets converts formInsets to NSDirectionalEdgeInsets using top,
left, and right. This is used when giving the top-most component padding
when there is some other element on the screen (e.g. a button at the
bottom), and the bottom-most component will handle bottom padding
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
https://jira.corp.stripe.com/browse/MOBILESDK-3609

https://docs.google.com/document/d/1twl3TcL0SN6G6AdyzBKzG5jK1KuBBmvw0xNCiBMN_Xk/edit?tab=t.0#heading=h.570u4leviueu
## Testing
<!-- How was the code tested? Be as specific as possible. -->
Updated snapshot tests, clicked through
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
[Added] The ability to customize form insets with
`PaymentSheetAppearance.formInsets`.