Add default opt-in for inline Link signup (#4937)
## Summary
<!-- Simple summary of what was changed. -->

This pull request adds default opt-in as a form of Link inline signup
for US consumers.

The functionality is currently behind a local feature flag and will also
be gated by a backend flag (`link_mobile_disable_default_opt_in`).

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Unit and snapshot tests.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->