Allow highlighting section borders on selection (#4935)
## Summary

This makes some changes to `SectionElement` and `SectionView` to support
showing borders around the element when it is selected. To do this, we
add a new configuration `SectionSelectionBehavior`, which we use to
style and animate the border on selection.

This PR also applies this configuration to the fields in the
`PayWithLinkSignUpViewController`.

## Motivation


https://www.figma.com/design/OGzRm5DunZ6ih4qHo73vEO/%F0%9F%8D%8F%F0%9F%93%B1-Mobile-Payment-Element?node-id=158-3864&t=Cj1NI1BnB2qFXKgl-0

## Testing


https://github.com/user-attachments/assets/ff7e2cf9-9f10-46a2-bc3b-5e7f8e9c8ec7

## Changelog

N/a