Klarna mandate double period (#4921)
## Summary
<!-- Simple summary of what was changed. -->
If the merchant display name ends with a period, leave it off so that we
don't get a double period.
If it ends with multiple periods or other punctuation, I think that's a
them problem, but it's reasonable for a merchant name to end with a
period (for example, Example, Inc.)
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-3562](https://jira.corp.stripe.com/browse/MOBILESDK-3562)
## Testing
<!-- How was the code tested? Be as specific as possible. -->
<img width="559" alt="Screenshot 2025-05-20 at 3 40 15 PM"
src="https://github.com/user-attachments/assets/aa25f567-4407-40b2-a6ba-9631145993c8"
/>
## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A