Fix incorrect primary button height for Link UI (#4967)
## Summary
<!-- Simple summary of what was changed. -->

This pull request makes sure that the Link primary button is rendered
with a fixed height, as opposed to being influenced by the new
`appearance.primaryButton.height` introduced in
https://github.com/stripe/stripe-ios/pull/4939.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Snapshot test.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->