Prevent Link dismissal more robustly (#4966)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an issue where the Link bottom sheet was
dismissible after a successful confirmation, but before the success
animation had finished. It also makes the sheet non-dismissible during
the signup and update calls.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

[LINK_MOBILE-7](https://jira.corp.stripe.com/browse/LINK_MOBILE-7)

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->