Limit prompting for Link in FlowController (#4977)
## Summary
<!-- Simple summary of what was changed. -->

This pull request limits prompting for Link in FlowController if the
user has previously rejected Link. This only applies when we attempt to
open the Link UI instead of FlowController, not if the user manually
selects Link in the list of payment options.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->