Remove fruit store demo from Payment Sheet Example app (#4972)
## Summary
The Fruit Store demo was temporarily added as a short term solution to
make it available quickly, and it is now available as intended as a
standalone app available via Testflight, so this cleans it up.

## Motivation
Maintaining two different versions of the app is challenging from a
backend perspective, onerous from a mobile perspective, and causes
universal link problems to keep both supported, and so it is better to
have everyone use the standalone app.

## Testing
Ensured that Fruit Store is no longer an option in the app and that
references to it no longer exist in the codebase.

## Changelog
N/A