Fix build on Xcode 26 (#4985)
## Summary
Remove this override, as it breaks Objective-C header generation due to
the deprecated method. It seems like the default answer is `true`, so
the code to open the URL still functions: I'm not sure why it was even
there.

## Motivation
Fix compiling on Xcode 26

## Testing
Locally in Xcode 26 beta and iOS 26

## Changelog
None