Show Link consumer email in row buttons (#4970)
## Summary
<!-- Simple summary of what was changed. -->

This pull request make the Link row button render the email address of a
returning Link user as the subtitle, instead of rendering a generic
message. We do this in both FlowController (in vertical mode) and
Embedded.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Snapshot tests.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->