Update secondary button usage in Link UI (#4994)
## Summary
<!-- Simple summary of what was changed. -->

This pull request updates the secondary button style and usage in Link
UI.

We no longer show the secondary button in the add and update screens,
and now use a plain style in the wallet screen.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

Snapshot tests.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->