Undo payment option label changes for Link (#5008)
## Summary
<!-- Simple summary of what was changed. -->

This pull request undoes a change where we handed the Link payment
method details to the merchant as part of `PaymentOption.label`.
Instead, we'll use a new construct `Labels` where we provide a label and
sublabel. This will come in a future pull request.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->