Enable Link RUX in FlowController (#4998)
## Summary
<!-- Simple summary of what was changed. -->

This pull request enables the Link RUX in FlowController. The feature is
still gated by a backend flag.

We changed two tests who were using Link as a stand-in for a payment
method that doesn't require any form input.

## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->