Initial Shop Pay WKWebView bridge (#5005)
## Summary
Integrates a one-off ECE bridge for Shop Pay. We'll eventually extend
this into a generic ECE bridge, but we'll start by hardcoding Shop Pay
support.

* ECEViewController presents ECE in a WKWebView and bridges the ECE
callbacks to the `ExpressCheckoutWebviewDelegate`.
* ShopPayECEPresenter uses ECEViewController to present Shop Pay,
including bridging the ECE callbacks to the user's shipping address/rate
handlers

This doesn't actually work end-to-end yet, as we haven't built out the
new facilitated payment confirmation handler. It'll also need a *lot* of
work if/when we enable it outside facilitated payment scenarios.

## Motivation
Enabling Shop Pay in WalletButtonsView

## Testing
Added integration tests, snapshot tests (fairly useless at the moment,
but we'll style it soon), and unit tests for the ViewController and JS
bridge.

## Changelog
N/A