Attempt to deflake: testLoaderStartShowsLoadingIndicator (#5019)
## Summary
This test seems flakey -- i suspect it is because we are asserting too
quickly. I wasn't able to repro but, the assertion failure was happening
on the line, and it's probable we are just asserting too quickly.

Example:
https://app.bitrise.io/build/6a2542b2-773d-43e7-99ad-b813330fe086

## Motivation
Flakey tests are annoying

## Testing
Run locally.

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->