Fix Link view controller jank (#5030)
## Summary
<!-- Simple summary of what was changed. -->

This pull request fixes an animation issue when opening the `Add payment
method` and `Update card` screens in the native Link UI, where the
bottom sheet height would suddenly change at the end of the animation.

### Before


https://github.com/user-attachments/assets/7a80b5d8-3b07-44df-9938-88121f8ae16b

### After


https://github.com/user-attachments/assets/1219b357-f5f7-4e8a-90e9-c183b2a20fcf


## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->

## Testing
<!-- How was the code tested? Be as specific as possible. -->

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->