Add support for scoped client secrets (#5041)
## Summary
* Adjust the PI client_secret regexes to support scoped client_secrets.
* When loading a scoped client secret, relax the usual validation and
inject redacted placeholders. We won't expose these scoped PIs to users,
so we don't want them to be optional fields in STPPaymentIntent, and
this seemed more straightforward and than making a more generic
PaymentIntent protocol.
* This leaves us without the client secret in the PI object, but we need
to reuse it later during 3DS2 processing to re-fetch the PI and check
the status. To work around this, we'll adjust `fetchPaymentIntent` to
inject the client secret. (I also looked into injecting this more
directly when decoding the original object, but the design of
STPAPIResponseDecodable makes that tricky. This approach is definitely
hacky, but much less disruptive.)

## Motivation
Support for scoped client secrets and next_actions.

## Testing
Added STPPaymentIntent tests for scoped PaymentIntent access, added an
end-to-end example in PaymentSheet Example.

## Changelog
N/A