Reduce shopPay jank (#5042)
## Summary
An attempt to remove jank
* Removes close button on the actual webview, replaced by a native
button. Couldn't figure out how to hook into this... -- open to other
ideas.
* Fixes memory leak by removing webview from controller
* Removes navigationController container to get rid of additional white
space at the top

## Motivation
Experience before was a bit janky.

## Testing
Manually tested -- and will need to update unit tests, if needed

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->