Assert loudly on integration errors with CustomerSession (#5053)
## Summary
Creating a customerSession with the wrong component will lead to
undefined behavior within our surfaces -- we should assert loudly before
continuing.

## Motivation
Currently if a user creates a customerSession with component
"payment_element" and passes it into PaymentSheet, it leads to silently
failing! This is bad! We should loudly fail instead!

## Testing
Manually tested by creating customerSessions with the wrong component,
and ensuring things fail!

## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->