traitCollectionDidChange loading indicator (#4928)
## Summary
<!-- Simple summary of what was changed. -->
The loading indicator color used to be based off of the system mode
(light or dark) instead of the component color, and it wouldn't change
its color in response to an appearance change. Now we use
traitCollectionDidChange to calculate the correct color and change it
when necessary.
## Motivation
<!-- Why are you making this change? If it's for fixing a bug, if
possible, please include a code snippet or example project that
demonstrates the issue. -->
[MOBILESDK-3426](https://jira.corp.stripe.com/browse/MOBILESDK-3426)
## Testing
<!-- How was the code tested? Be as specific as possible. -->
Before


https://github.com/user-attachments/assets/67209eed-8fb0-4711-bc4a-db3c257385f2

After



https://github.com/user-attachments/assets/f655440a-c0e6-4857-922f-4f4e4dcb9518



## Changelog
<!-- Is this a notable change that affects users? If so, add a line to
`CHANGELOG.md` and prefix the line with one of the following:
    - [Added] for new features.
    - [Changed] for changes in existing functionality.
    - [Deprecated] for soon-to-be removed features.
    - [Removed] for now removed features.
    - [Fixed] for any bug fixes.
    - [Security] in case of vulnerabilities.
-->
N/A